Improving packaging structure
