+++ app/src/main/AndroidManifest.xml
@@ -25,7 +25,7 @@
         android:required="false"/>
 
     <application
-        android:name=".app.BrowserApp"
+        android:name=".BrowserApp"
         android:allowBackup="true"
         android:hardwareAccelerated="true"
         android:icon="@mipmap/ic_launcher"
@@ -38,7 +38,7 @@
             android:value="2.1"/>
 
         <activity
-            android:name=".activity.MainActivity"
+            android:name=".MainActivity"
             android:alwaysRetainTaskState="true"
             android:configChanges="orientation|screenSize|screenLayout|smallestScreenSize|keyboardHidden|keyboard"
             android:label="@string/app_name"
@@ -122,7 +122,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".activity.SettingsActivity"
+            android:name=".settings.activity.SettingsActivity"
             android:configChanges="orientation|screenSize|screenLayout|smallestScreenSize|keyboardHidden|keyboard"
             android:label="@string/settings"
             android:theme="@style/Theme.SettingsTheme">
@@ -133,7 +133,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".activity.IncognitoActivity"
+            android:name=".IncognitoActivity"
             android:alwaysRetainTaskState="true"
             android:configChanges="orientation|screenSize|screenLayout|smallestScreenSize|keyboardHidden|keyboard"
             android:label="@string/app_name"
@@ -147,7 +147,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".activity.ReadingActivity"
+            android:name=".reading.activity.ReadingActivity"
             android:configChanges="orientation|screenSize|screenLayout|smallestScreenSize|keyboardHidden|keyboard"
             android:label="@string/reading_mode"
             android:theme="@style/Theme.SettingsTheme">
