+++ app/src/main/java/acr/browser/lightning/BrowserApp.java
@@ -1,4 +1,4 @@
-package acr.browser.lightning.app;
+package acr.browser.lightning;
 
 import android.app.Activity;
 import android.app.Application;
@@ -20,11 +20,13 @@ import java.util.List;
 
 import javax.inject.Inject;
 
-import acr.browser.lightning.BuildConfig;
 import acr.browser.lightning.database.HistoryItem;
 import acr.browser.lightning.database.bookmark.BookmarkExporter;
-import acr.browser.lightning.database.bookmark.legacy.LegacyBookmarkManager;
 import acr.browser.lightning.database.bookmark.BookmarkModel;
+import acr.browser.lightning.database.bookmark.legacy.LegacyBookmarkManager;
+import acr.browser.lightning.di.AppComponent;
+import acr.browser.lightning.di.AppModule;
+import acr.browser.lightning.di.DaggerAppComponent;
 import acr.browser.lightning.preference.PreferenceManager;
 import acr.browser.lightning.utils.FileUtils;
 import acr.browser.lightning.utils.MemoryLeakUtils;
