+++ app/src/main/java/acr/browser/lightning/browser/SearchBoxModel.java
@@ -8,7 +8,7 @@ import android.text.TextUtils;
 import javax.inject.Inject;
 
 import acr.browser.lightning.R;
-import acr.browser.lightning.app.BrowserApp;
+import acr.browser.lightning.BrowserApp;
 import acr.browser.lightning.preference.PreferenceManager;
 import acr.browser.lightning.utils.UrlUtils;
 import acr.browser.lightning.utils.Utils;
