+++ app/src/main/java/acr/browser/lightning/browser/bookmark/BookmarkUiModel.java
@@ -1,4 +1,4 @@
-package acr.browser.lightning.activity;
+package acr.browser.lightning.browser.bookmark;
 
 import android.support.annotation.Nullable;
