+++ app/src/main/java/acr/browser/lightning/constant/HistoryPage.java
@@ -25,7 +25,7 @@ import java.util.List;
 import javax.inject.Inject;
 
 import acr.browser.lightning.R;
-import acr.browser.lightning.app.BrowserApp;
+import acr.browser.lightning.BrowserApp;
 import acr.browser.lightning.database.HistoryItem;
 import acr.browser.lightning.database.history.HistoryModel;
 import acr.browser.lightning.utils.Preconditions;
