+++ app/src/main/java/acr/browser/lightning/constant/StartPage.java
@@ -17,7 +17,7 @@ import java.io.IOException;
 import javax.inject.Inject;
 
 import acr.browser.lightning.R;
-import acr.browser.lightning.app.BrowserApp;
+import acr.browser.lightning.BrowserApp;
 import acr.browser.lightning.search.SearchEngineProvider;
 import acr.browser.lightning.search.engine.BaseSearchEngine;
 import acr.browser.lightning.utils.Utils;
