+++ app/src/main/java/acr/browser/lightning/di/AppComponent.java
@@ -1,12 +1,13 @@
-package acr.browser.lightning.app;
+package acr.browser.lightning.di;
 
 import javax.inject.Singleton;
 
-import acr.browser.lightning.activity.BrowserActivity;
-import acr.browser.lightning.activity.ReadingActivity;
-import acr.browser.lightning.activity.TabsManager;
-import acr.browser.lightning.activity.ThemableBrowserActivity;
-import acr.browser.lightning.activity.ThemableSettingsActivity;
+import acr.browser.lightning.browser.activity.BrowserActivity;
+import acr.browser.lightning.reading.activity.ReadingActivity;
+import acr.browser.lightning.browser.TabsManager;
+import acr.browser.lightning.browser.activity.ThemableBrowserActivity;
+import acr.browser.lightning.settings.activity.ThemableSettingsActivity;
+import acr.browser.lightning.BrowserApp;
 import acr.browser.lightning.browser.BrowserPresenter;
 import acr.browser.lightning.browser.SearchBoxModel;
 import acr.browser.lightning.constant.BookmarkPage;
@@ -16,13 +17,13 @@ import acr.browser.lightning.constant.StartPage;
 import acr.browser.lightning.dialog.LightningDialogBuilder;
 import acr.browser.lightning.download.DownloadHandler;
 import acr.browser.lightning.download.LightningDownloadListener;
-import acr.browser.lightning.fragment.BookmarkSettingsFragment;
-import acr.browser.lightning.fragment.BookmarksFragment;
-import acr.browser.lightning.fragment.DebugSettingsFragment;
-import acr.browser.lightning.fragment.GeneralSettingsFragment;
-import acr.browser.lightning.fragment.LightningPreferenceFragment;
-import acr.browser.lightning.fragment.PrivacySettingsFragment;
-import acr.browser.lightning.fragment.TabsFragment;
+import acr.browser.lightning.settings.fragment.BookmarkSettingsFragment;
+import acr.browser.lightning.browser.fragment.BookmarksFragment;
+import acr.browser.lightning.settings.fragment.DebugSettingsFragment;
+import acr.browser.lightning.settings.fragment.GeneralSettingsFragment;
+import acr.browser.lightning.settings.fragment.LightningPreferenceFragment;
+import acr.browser.lightning.settings.fragment.PrivacySettingsFragment;
+import acr.browser.lightning.browser.fragment.TabsFragment;
 import acr.browser.lightning.search.SearchEngineProvider;
 import acr.browser.lightning.search.SuggestionsAdapter;
 import acr.browser.lightning.utils.ProxyUtils;
