+++ app/src/main/java/acr/browser/lightning/dialog/LightningDialogBuilder.java
@@ -22,8 +22,8 @@ import java.util.List;
 import javax.inject.Inject;
 
 import acr.browser.lightning.R;
-import acr.browser.lightning.activity.MainActivity;
-import acr.browser.lightning.app.BrowserApp;
+import acr.browser.lightning.MainActivity;
+import acr.browser.lightning.BrowserApp;
 import acr.browser.lightning.constant.BookmarkPage;
 import acr.browser.lightning.constant.Constants;
 import acr.browser.lightning.controller.UIController;
