+++ app/src/main/java/acr/browser/lightning/download/DownloadHandler.java
@@ -32,8 +32,8 @@ import javax.inject.Inject;
 
 import acr.browser.lightning.BuildConfig;
 import acr.browser.lightning.R;
-import acr.browser.lightning.activity.MainActivity;
-import acr.browser.lightning.app.BrowserApp;
+import acr.browser.lightning.MainActivity;
+import acr.browser.lightning.BrowserApp;
 import acr.browser.lightning.constant.Constants;
 import acr.browser.lightning.controller.UIController;
 import acr.browser.lightning.database.downloads.DownloadItem;
