+++ app/src/main/java/acr/browser/lightning/utils/Utils.java
@@ -44,7 +44,7 @@ import java.text.SimpleDateFormat;
 import java.util.Date;
 
 import acr.browser.lightning.R;
-import acr.browser.lightning.activity.MainActivity;
+import acr.browser.lightning.MainActivity;
 import acr.browser.lightning.constant.Constants;
 import acr.browser.lightning.database.HistoryItem;
 import acr.browser.lightning.dialog.BrowserDialog;
