+++ app/src/main/java/acr/browser/lightning/view/LightningChromeClient.java
@@ -25,7 +25,7 @@ import com.anthonycr.grant.PermissionsResultAction;
 import javax.inject.Inject;
 
 import acr.browser.lightning.R;
-import acr.browser.lightning.app.BrowserApp;
+import acr.browser.lightning.BrowserApp;
 import acr.browser.lightning.controller.UIController;
 import acr.browser.lightning.dialog.BrowserDialog;
 import acr.browser.lightning.favicon.FaviconModel;
