+++ app/src/main/java/acr/browser/lightning/view/LightningView.java
@@ -44,7 +44,7 @@ import java.util.Map;
 
 import javax.inject.Inject;
 
-import acr.browser.lightning.app.BrowserApp;
+import acr.browser.lightning.BrowserApp;
 import acr.browser.lightning.constant.BookmarkPage;
 import acr.browser.lightning.constant.Constants;
 import acr.browser.lightning.constant.DownloadsPage;
